+++ payloads/v1/index.ts
@@ -4,4 +4,3 @@ export * from './emote';
 export * from './events';
 export * from './oauth2';
 export * from './users';
-export * from './webhooks';
